+++ CHANGELOG.md
@@ -6,15 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.1.2] - 2024-11-13xxxxxxxxx
+## [0.2.0] - 2024-11-17
+- breaking change, aligned function names with HX711 where possible 
 - implement base class HX710AB (reduce source duplication)
 - implement async interface, three functions
-  - **void request()**, **bool isReady()**, **fetch()**
-- implement two point calibration **void calibrateUnit()**
+  - **void request()**, **bool is_ready()**, **fetch()**
+- implement two point calibration **void calibrate()**
 - add unit tests
 - add examples
 - update keywords.txt
 - update readme.md
+
+----
 
 ## [0.1.1] - 2024-11-13
 - fix #2, add fastProcessor flag.